remove webmock dependency and associated test macro
prefer configuring the signature verifier strategy.
