+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in auth-rb.gemspec
+# Specify your gem's dependencies in keratin-authn.gemspec
 gemspec
