+++ test/keratin_authn_test.rb
@@ -131,4 +131,23 @@ class Keratin::AuthNTest < Keratin::AuthN::TestCase
       exp: (Time.now + 150).to_i
     }
   end
+
+  private def stub_auth_server(issuer: Keratin::AuthN.config.issuer, keypair: jws_keypair)
+    Keratin::AuthN.signature_verifier.keychain.clear
+    stub_request(:get, "#{issuer}/configuration").to_return(
+      status: 200,
+      body: {'jwks_uri' => "#{issuer}/jwks"}.to_json
+    )
+    stub_request(:get, "#{issuer}/jwks").to_return(
+      status: 200,
+      body: {
+        keys: [
+          keypair.public_key.to_jwk.slice(:kty, :kid, :e, :n).merge(
+            use: 'sig',
+            alg: JWS_ALGORITHM
+          )
+        ]
+      }.to_json
+    )
+  end
 end
